flatten split files
