+++ packages/react-native-gesture-handler/android/src/main/java/com/swmansion/gesturehandler/ReactContextExtensions.kt
